restore legacy branch so the original port can be played without my rendering rework

+++ BlockByBlock/net/minecraft/src/NBTTagByteArray.cs
@@ -46,7 +46,7 @@ namespace net.minecraft.src
 		{
 			get
 			{
-				return 7;
+				return (sbyte)7;
 			}
 		}
